requirements.txt
streamlit
faiss-cpu
numpy
sentence-transformers
openai
PyPDF2
pdf2image
pytesseract
Pillow
torch
google-api-python-client
google-auth
groq
